backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
mysql-connector-python==8.4.0
groq==0.9.0
python-dotenv